重构人员信息页面

+++ backend/backend/src/main/java/top/shanhai1024/repository/PersonRepository.java
@@ -3,5 +3,5 @@ package top.shanhai1024.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import top.shanhai1024.entity.PO.Person;
 
-public interface PersonRepository  extends JpaRepository<Person,Integer> {
+public interface PersonRepository  extends JpaRepository<Person,Long> {
 }
